update on task 6

+++ api/v1/views/states.py
@@ -19,7 +19,9 @@ def get_states():
                  methods=["GET"], strict_slashes=False)
 def get_state(state_id):
     """Retrieve a specific State object"""
-    state = storage.get_or_404(State, state_id)
+    state = storage.get(State, state_id)
+    if state is None:
+        abort(404)
     return jsonify(state.to_dict())
 
 
@@ -27,7 +29,9 @@ def get_state(state_id):
                  methods=["DELETE"], strict_slashes=False)
 def delete_state(state_id):
     """Delete a State object"""
-    state = storage.get_or_404(State, state_id)
+    state = storage.get(State, state_id)
+    if state is None:
+        abort(404)
     storage.delete(state)
     storage.save()
     return jsonify({}), 200
@@ -52,7 +56,9 @@ def create_state():
                  methods=["PUT"], strict_slashes=False)
 def update_state(state_id):
     """Update an existing State object"""
-    state = storage.get_or_404(State, state_id)
+    state = storage.get(State, state_id)
+    if state is None:
+        abort(404)
     data = request.get_json()
     if not data:
         abort(400, "Not a JSON")
